Image Enhancement/requirements.txt
matlab==0.1
matplotlib==3.3.2
numpy==1.19.2
opencv_contrib_python==4.5.1.48
Pillow==9.2.0
